Added any_character_set and the character<any_character_set> specialization.
This change adds ID tagging for all character sets so that characters
associated with the any_character_set can be dynamically associated with other
character sets.

Conversion constructors and operators for character/code point conversions
were removed in favor of explicit getters and setters.  This exposed the fact
that the Character concept required an ordered type, but that the character
class was not supplying the necessary operators.  Previously, the ordering
requirements were being met via implicit conversion to code point types.
This was resolved by removing ordering requirements for characters.