Changed the encoding concept to require modeled types to have an initial_state static member function rather than a static data member.
This change was made to allow the address of the initial state object to be
taken without having to define that object in a particular TU.  For example,
an encoding can define an inline initial_state() member function containing
a static state object to which a reference is returned.  Linkage rules will
ensure that only one such object is constructed.