Addressed the FIXME issue regarding compilation failures initializing codec state base classes.
List-initialization of aggregates does not consider copy constructors.
This is the subject matter of CWG DR1467:
  http://www.open-std.org/jtc1/sc22/wg21/docs/cwg_defects.html#1467
Gcc trunk (5.0, c++-concepts branch, svn revision 219158) contains a
partial implementation of the proposed resolution, but fails to address
initialization of an aggregate used as a base class.  Implementation
of DR1467 was performed via gcc bug 51747:
  https://gcc.gnu.org/bugzilla/show_bug.cgi?id=51747
This bug has been reopened to follow up on use of list-initialization
to invoke a copy constructor for aggregate base classes.  In the mean
time, an expression-list must be used in lieu of list-initialization.