Modified codec decode() and rdecode() to support decoding non-code-point encoding segments.
Previously, every call to decode() or rdecode() was required to return
a decoded code point.  For stateful codecs, this made it difficult to
unambiguously associate a code unit sequence with a code point since
a code unit sequence could encode state transitions that are not
strictly part of the encoding of a code point.   Associating such
sequences with a code point would violate assumptions with respect to
max_code_units.

This change required modifying itext_iterator to loop on calls to
decode() and rdecode() until the end of the range is reached or until
a code point is decoded.

encode() and rencode() were not similarly updated, so this change breaks
symmetry with decode() and rdecode().  They still return void and, for
stateful encodings, may write more than max_code_units code units when
called for a single code point if a state encoding code unit sequence
is required.