Modified itext_iterator to determine the underlying code unit iterator using a const qualified range type.
Previously, the underlying code unit iterator of an itext_iterator was
identified using Iterator_type on the range type template argument.  This
resulted in constructors that could not be constructed with const_iterator
types unless the range type was itself const qualified.  Now, the range
type is explicitly const qualified so as to obtain a const_iterator type.