Added range_based_for.hpp and the RANGE_BASED_FOR macro based on Eric Niebler's RANGES_FOR macro in his ranges-v3 library.
The C++11 range-based-for statement requires that the begin and end iterator
types for the container type be identical types.  Since text_view uses sentinel
types for the end type, it cannot be used with the C++11 range-based-for.  The
RANGE_BASED_FOR macro provides an alternate implementation that does work with
text_view.  This macro will hopefully be a temporary workaround made obsolete
in C++17.