Removed the codec related concepts and archetypes
The codec related concepts turned out not to be useful for typical
use cases and, worse, complicated typical use cases by requiring
unintuitive use of member types nested in codec types.  With this
change, the functionality previously exposed via codecs is now
exposed directly in encoding objects.  The previous codec types
are still present, but are now an implementation detail.