Removed make_cstr_view() from the text_view header and specification.
make_cstr_view() was only used by test utilities, has a poor name
(it sounds like it should return a view of a string *with* a null
terminator), and is rather error prone to use (it doesn't actually
validate that a string terminator was present).  A more suitable
replacement, should a need to construct views that omit/append a
string terminator persist, would be several functions that return
views that lazily omit/append terminators.