Replaced uses of RANGE_BASED_FOR with the C++17 range-based for statement; removed range_based_for.hpp.
P0184R0 was adopted for C++17 and implementation support was added to
gcc in r234191 (though the corresponding feature test macro was not
updated until r234508).  This change removes previous workarounds that
were in place pending relaxation of the exact type requirements for
the begin and end expressions of the C++11 range-based for statement.