Migrated to a cmake based build system.
Tristan Brindle was kind enough to contribute a basic cmake based
build system.  This change expands on that initial contribution and
removes the previous 'make' based build system.  CMake 3.0.2 or later
is now required to build the Text_view tests, validate the example
programs, and perform an installation.

Included with this change are:
- a CMake 'find' module for CMCSTL2.  The location of the CMCSTL2
  installation/repo is still expected to be provided by setting the
  CMCSTL2_INSTALL_PATH environment variable.
- Renaming and relocation of the example code.  Dashes are now
  preferred over underscores in most places.  Header files retain
  their preference for underscores for consistency with standard
  C++ headers.  Each example program now has a dedicated directory.
- Installation is now supported.  The installation consists of header
  files, example code, a license file, and the readme file.  The
  text_view library is still header-only, so the installation can
  still be skipped.