CMake updates to enable modular build components.
This change enables the examples to be built both in-tree and from an
installation directory using the same CMake files.  A text_view package
find module is provided for in-tree use (only) and it sets additional
compiler options to ensure the example code remains warning free.  The
installed text_view config module allows more permissive compilation.

This change also fixes some issues with the CMCSTL2 package find module
to ensure it works as expected when CMake files in different directories
invoke find_package for it.