Corrected some iterator invalidation issues with wrapping of input and output iterators in itext_iterator and otext_iterator.
Previously, itext_cursor and otext_cursor attempted to meet exception guarantees
by copying their stored iterators for encode and decode operations so that a
thrown exception would leave the stored iterators unmutated.  This worked for
forward or better iterators, but resulted in undefined behavior for input and
output iterators since it was possible for the copy to be mutated but not
assigned back if an exception was thrown.  This fixes issue #28.