Updated TextOutputIterator to refine TextIterator.
Previously, the TextOutputIterator concept copied most of the requirements
of TextIterator.  The reason it wasn't a refinement was because TextIterator
included a requirement that value type satisfy Character; that requirement
is not present for TextOutputIterator.  With this change, the requirement
that the iterator value type satisfy Character has been moved to
TextInputIterator such that TextOutputIterator now properly refines
TextIterator.

This change also updates the make_text_view() overload that took TextIterator
parameters to now require TextInputIterator.